Removed Module Descriptions from all projects.
Updated all projects to work with SSP version 1.1.0.

Added ADC source code

+++ Lesson004/synergy/ssp/src/bsp/mcu/s7g2/bsp_register_protection.c
@@ -155,9 +155,9 @@ void bsp_register_protect_open (void)
     uint32_t i;
 
     /** Initialize reference counters to 0. */
-    for (i = 0; i < BSP_REG_PROTECT_TOTAL_ITEMS; i++)
+    for (i = (uint32_t)0; i < BSP_REG_PROTECT_TOTAL_ITEMS; i++)
     {
-        g_protect_counters[i] = 0;
+        g_protect_counters[i] = (uint16_t)0;
     }
 }
